Worked on the TSE recommendation
